refactor(client): delete MinimapRenderer.ts (-941 lines), extract minimapTypes.ts
MinimapRenderer.ts is fully replaced by the shared MinimapWorker. All canvas
drawing (roads, buildings, entity pips, POI icons, destination marker) now
runs off-thread in the worker.

Remaining types (MinimapRenderState, HyperscapeWindow, createRenderState,
getSpectatorTarget) extracted to minimapTypes.ts for use by Minimap.tsx
and useMinimapInteraction.ts.
